Revert "Add google instance to GoogleLoader params"
This reverts commit 63b251f51334f34dc6a0b4582523a451edd34319.
