+++ adsmodule/src/main/java/tv/limehd/adsmodule/google/Google.kt
@@ -1,7 +1,6 @@
 package tv.limehd.adsmodule.google
 
 import android.content.Context
-import android.os.Handler
 import android.util.Log
 import tv.limehd.adsmodule.LimeAds
 import tv.limehd.adsmodule.interfaces.AdRequestListener
@@ -27,41 +26,13 @@ class Google(private val context: Context,
         private const val TAG = "Google"
     }
 
-    //********************************************* GOOGLE INTERSTITIAL TIMER HANDLER ****************************************************** //
-
-    val googleTimerHandler: Handler = Handler()
-    var timer = 30
-    var isAllowedToRequestGoogleAd = true
-    val googleTimerRunnable: Runnable = object : Runnable {
-        override fun run() {
-            if (timer > 0) {
-                timer--
-                Log.d(TAG, "Google timer: $timer")
-                googleTimerHandler.postDelayed(this, 1000)
-            }else{
-                isAllowedToRequestGoogleAd = true
-            }
-        }
-    }
-
     /**
-     * Get Google Ad
-     * Check if we are allowed to load google interstitial ad, cause of the timer
-     *
-     * @param   isLoadInterstitial  stands for understanding are we loading interstitial or in main stream with all ads
+     * Получить рекламу для площадки Google
      */
 
     fun getGoogleAd(isLoadInterstitial: Boolean) {
         Log.d(TAG, "Load google ad")
-        val googleLoader = GoogleLoader(context, lastAd, fragmentState, adRequestListener, adShowListener, isLoadInterstitial, limeAds, this)
-        if(isLoadInterstitial) {
-            if (isAllowedToRequestGoogleAd) {
-                isAllowedToRequestGoogleAd = false
-                timer = 0
-                googleLoader.loadAd()
-            }
-        }else{
-            googleLoader.loadAd()
-        }
+        val googleLoader = GoogleLoader(context, lastAd, fragmentState, adRequestListener, adShowListener, isLoadInterstitial, limeAds)
+        googleLoader.loadAd()
     }
 }
